pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evaluating-epilink"
version = "0.1.0"
description = "Reproducibility workflows and analyses for the epilink manuscript."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Dominic Arthur" }
]
dependencies = [
  "bcubed",
  "igraph",
  "matplotlib",
  "networkx",
  "numpy",
  "pandas",
  "pyarrow",
  "pyyaml",
  "scikit-learn",
  "scipy",
  "seaborn",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
